// app/src/types/EscUserResult.ts
export interface EscUserResult {
    entry_id: number;
    country: string;
    flag_url: string;
    artist: string;
    song_title: string;
    song_avg: number;
    costume_avg: number;
    staging_avg: number;
    performance_avg: number;
    total_avg: number;
    place: number;
    real_place: number;
    diff: number;
}